requirements.txt
asgiref==3.2.7
backcall==0.1.0
certifi==2020.4.5.1
chardet==3.0.4
decorator==4.4.2
Django==3.0.5
django-cors-headers==3.2.1
django-rest-passwordreset==1.1.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==0.4.1
httpie==2.1.0
idna==2.9
ipdb==0.13.2
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.17.0
mysqlclient==1.4.6
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.1
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
PyJWT==1.7.1
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.1.9
